creates controller package and refactors
