+++ src/main/java/de/rabea/Main.java
@@ -1,6 +1,9 @@
 package de.rabea;
 
 import de.rabea.controller.*;
+import de.rabea.logger.ConsoleLogger;
+import de.rabea.logger.FileLogger;
+import de.rabea.logger.MultiLogger;
 import de.rabea.request.Directory;
 
 import java.io.File;
@@ -9,6 +12,8 @@ import java.net.ServerSocket;
 import java.util.concurrent.Executors;
 
 public class Main {
+    private static final String logFilePath = new File("../../logs.txt").getAbsolutePath();
+
     public static void main(String[] args) throws IOException {
         Arguments arguments = new Arguments(args);
         int port = arguments.getPort();
@@ -17,8 +22,10 @@ public class Main {
 
         ServerSocket serverSocket = new ServerSocket(port);
         Router router = new Router();
+
         MultiLogger logger = new MultiLogger();
-        logger.add(new FileLogger(new File("../../logs.txt").getAbsolutePath()), new ConsoleLogger(new ServerConsole()));
+        logger.add(new FileLogger(logFilePath), new ConsoleLogger(new ServerConsole()));
+
         router.configure("/", new RootController(directory));
         router.configure("/coffee", new CoffeeController());
         router.configure("/logs", new LogsController(logger));
@@ -29,6 +36,7 @@ public class Main {
         router.configure("/redirect", new RedirectController());
         router.configure("/tea", new TeaController());
         router.configure(directory, new AssetController(directory));
+
         HttpServer httpServer = new HttpServer(Executors.newFixedThreadPool(20), serverSocket, router, logger);
         httpServer.start(givenDirectory, port);
     }
