+++ src/main/java/de/rabea/logger/ConsoleLogger.java
@@ -1,4 +1,6 @@
-package de.rabea;
+package de.rabea.logger;
+
+import de.rabea.Console;
 
 public class ConsoleLogger implements Logger {
