+++ src/main/java/de/rabea/logger/Logger.java
@@ -1,4 +1,4 @@
-package de.rabea;
+package de.rabea.logger;
 
 public interface Logger {
     void log(String message);
