+++ src/test/java/de/rabea/FakeConsole.java
